+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPX4DPK0MYYK66VPF6HZ5RX2"
+        "01KPX8CSJ7WDKB608J0NKDFFY3"
     )
     r"""The unique identifier of the data source"""